Creates Playlist Controller. Adds index to playlist controller.
