+++ server.js
@@ -1,6 +1,7 @@
 pry = require('pryjs')
 
-const SongsController = require('./controllers/api/v1/songs_controller')
+const SongsController    = require('./controllers/api/v1/songs_controller')
+const PlaylistController = require('./controllers/api/v1/playlists_controller')
 const Song       = require('./lib/models/song');
 const Playlist   = require('./lib/models/playlist');
 const express    = require('express');
@@ -30,36 +31,35 @@ app.patch('/api/v1/songs/:id', SongsController.update);
 
 app.delete('/api/v1/songs/:id', SongsController.delete);
 
-app.post('/api/v1/playlists/:playlist_id/songs/:id', (request, response) => {
-  const playlist      = request.params.playlist_id
-  const song          = request.params.id
-  const playlistSong = {song_id: song, playlist_id: playlist}
+// app.post('/api/v1/playlists/:playlist_id/songs/:id', 
+//   const playlist = request.params.playlist_id
+//   const song = request.params.id
+//   const playlistSong = { song_id: song, playlist_id: playlist }
 
-  let songName     = ""
-  let playlistName = ""
-  database('songs')
-  .select('name')
-  .where('id', song)
-  .then(song => {
-    songName = song[0].name
-  })
+//   let songName = ""
+//   let playlistName = ""
+//   database('songs')
+//     .select('name')
+//     .where('id', song)
+//     .then(song => {
+//       songName = song[0].name
+//     })
 
-  database('playlists')
-  .select('name')
-  .where('id', playlist)
-  .then(playlist => {
-    playlistName = playlist[0].name
-  })
+//   database('playlists')
+//     .select('name')
+//     .where('id', playlist)
+//     .then(playlist => {
+//       playlistName = playlist[0].name
+//     })
 
-  database('playlist_songs')
-  .insert(playlistSong, ['song_id', 'playlist_id'])
-  .then(() => {
-    response.status(201).json({"message": `Successfully added ${songName} to ${playlistName}`});
-  })
-  .catch((error) => {
-    response.status(500).json({ error });
-  });
-});
+//   database('playlist_songs')
+//     .insert(playlistSong, ['song_id', 'playlist_id'])
+//     .then(() => {
+//       response.status(201).json({ "message": `Successfully added ${songName} to ${playlistName}` });
+//     })
+//     .catch((error) => {
+//       response.status(500).json({ error });
+//     }););
 
 app.delete('/api/v1/playlists/:playlist_id/songs/:id', (request, response) => {
   const playlist      = request.params.playlist_id
@@ -92,28 +92,7 @@ app.delete('/api/v1/playlists/:playlist_id/songs/:id', (request, response) => {
     })
 });
 
-app.get('/api/v1/playlists', (request, response) => {
-  let playlists = []
-  let songs = []
-
-  Playlist.all()
-  .then((a) => {
-    playlists = a
-  })
-
-  Song.withPlaylistId()
-  .then((a) => { songs = a })
-  .then(() => {
-    for(let playlist of playlists) {
-      playlist.songs = songs.filter(song => (song.playlist_id == playlist.id))
-      playlist.songs.forEach(song => delete song.playlist_id)
-    }
-  })
-  .then(() => {response.status(200).json(playlists)})
-  .catch((error) => {
-    response.status(500).json({ error });
-  });
-});
+app.get('/api/v1/playlists', PlaylistController.index);
 
 app.get('/api/v1/playlists/:playlist_id/songs', (request, response) => {
   let playlistId = request.params.playlist_id
